Added BigInteger support Added Exponentiation Operation

+++ EquationFactories/AlgebraicString/AlgebraicString.cs
@@ -1,28 +1,28 @@
 ﻿/*
  *
- * Developed by Adam Rakaska
- *  http://www.csharpprogramming.tips
+ * Developed by Adam White
+ *  https://csharpcodewhisperer.blogspot.com
  * 
  */
 using System;
 using System.Linq;
-using System.Text;
-using EquationFinderCore;
+using System.Numerics;
 using System.Globalization;
 using System.Collections.Generic;
+using EquationFinderCore;
 
 namespace EquationFactories
 {
 	public partial class AlgebraicString : IEquation
 	{
 		public bool IsCorrect { get { return (Result == TargetValue); } }
-		public decimal Result { get { return Solve(); } }
-		private decimal? _result = null;
+		public BigInteger Result { get { return Solve(); } }
+		private BigInteger? _result = null;
 		private string Equation { get; set; }
 		private IEquationFinderArgs EquationArgs { get; set; }
 		private List<int> TermPool { get { return EquationArgs.TermPool; } }
 		private string OperatorPool { get { return EquationArgs.OperatorPool; } }
-		private decimal TargetValue { get { return EquationArgs.TargetValue; } }
+		private BigInteger TargetValue { get { return EquationArgs.TargetValue; } }
 		private int NumberOfOperations { get { return EquationArgs.NumberOfOperations; } }
 
 		public AlgebraicString()
@@ -46,14 +46,14 @@ namespace EquationFactories
 			get { return (Result == TargetValue); }
 		}
 
-		private decimal Solve()
+		private BigInteger Solve()
 		{
 			if (_result == null)
 			{
 				_result = InfixNotation.Evaluate(Equation);
 				//_result = StaticScriptControl.Evaluate(Equation);
 			}
-			return (decimal)_result;
+			return (BigInteger)_result;
 		}
 
 		public override string ToString()
